Add files via upload
This is an exciting game project called - The Simon Game Project.  
It was a very challenging javascript project to work upon. The webpage is simple with basic html and css, but it has quite standard level of Javascript.

How to play the game?
The user is given a color, he has to click that color. In the next step he is given another color and this time he has click the new color following the previous color, and so on.